Added libraries to git and continued to edit the UI
Finished the Select / Create task group, as well as the file management
group. Just started the Schedule group, and so I added task scheduler
libraries to make the ui easier to manage.

Also added the libraries to the git so file dependencies don't get
disrupted when editing on a different machine